major refactor : -split code in modules -remove rename operator(batch rename by_default) -using obj exporter and importer -change variable to make it work for V-HACD 4.0
